// services/authentication-service/src/modules/auth/models/model-property-description.enum.ts
// Copyright (c) 2023 Sourcefuse Technologies
//
// This software is released under the MIT License.
// https://opensource.org/licenses/MIT
export const enum ModelPropertyDescriptionString {
  reqStrPropDesc = 'This property is supposed to be a string and is a required field',
  reqNumPropDesc = 'This property is supposed to be a number and is a required field',
}